Breakout components into separate files
 Extract TableViewHeader

Breakout TableViewFooter

Breakout TableViewCelll

Breakout TableViewItem

Breakout consts and tableview
