+++ src/TableViewConsts.js
@@ -1,0 +1,5 @@
+import { NativeModules } from 'react-native'
+
+const RNTableViewConsts = NativeModules.RNTableViewManager.Constants
+
+export default RNTableViewConsts
